+++ src/Engine/Assets/Shaders/Renderer_MeshDefaultShader.vert
@@ -1,6 +1,7 @@
 #version 460
 
 #extension GL_GOOGLE_include_directive : require
+#extension GL_ARB_gpu_shader_int64 : require
 
 #include "InputStructures.glsl"
 
@@ -22,12 +23,12 @@ layout(location = 3) in float uv_y;
 layout(location = 4) in vec4 color;
 
 layout (instanced location = 5) in mat4 model;
-layout (instanced location = 9) in int entityID;
+layout (instanced location = 9) in uint64_t entityID;
 
 layout(location = 0) out vec3 outNormal;
 layout(location = 1) out vec3 outColor;
 layout(location = 2) out vec2 outUV;
-layout(location = 3) out int outEntityID;
+layout(location = 3) out uint64_t outEntityID;
 
 void main()
 {
